V 2.8.12 Refactor
