+++ src/grpc_server/launcher.rs
@@ -5,236 +5,18 @@ use sysinfo::System;
 use tonic::{Request, Response, Status};
 use tracing::trace;
 
-use proto::shard_server::Shard;
-
 use crate::constant::{
     ACTIVITY_NAME, APP_VERSION, BOT_COMMANDS, BOT_INFO, GRPC_CERT_PATH, GRPC_KEY_PATH,
     GRPC_SERVER_PORT, GRPC_USE_TLS,
 };
 use crate::grpc_server::command_list::{
-    get_command_list, Arg, Command, CommandItem, SubCommand, SubCommandGroup,
+    get_list_of_all_command, Arg, Command, CommandItem, SubCommand, SubCommandGroup,
 };
-use crate::grpc_server::launcher::proto::command_service_server::{
-    CommandService, CommandServiceServer,
-};
-use crate::grpc_server::launcher::proto::info_server::{Info, InfoServer};
-use crate::grpc_server::launcher::proto::shard_server::ShardServer;
-use crate::grpc_server::launcher::proto::{
-    BotInfoData, CommandListRequest, CommandListResponse, InfoRequest, InfoResponse, SystemInfoData,
-};
+use crate::grpc_server::service;
+use crate::grpc_server::service::command::{CommandServices, get_command_server};
+use crate::grpc_server::service::info::{get_info_server, InfoService};
+use crate::grpc_server::service::shard::{get_shard_server, ShardService};
 
-// Proto module contains the protobuf definitions for the shard service
-mod proto {
-    // Include the protobuf definitions for the shard service
-    tonic::include_proto!("shard");
-    tonic::include_proto!("info");
-    tonic::include_proto!("command");
-    // FILE_DESCRIPTOR_SET is a constant byte array that contains the file descriptor set for the shard service
-    pub(crate) const SHARD_FILE_DESCRIPTOR_SET: &[u8] =
-        tonic::include_file_descriptor_set!("shard_descriptor");
-    pub(crate) const INFO_FILE_DESCRIPTOR_SET: &[u8] =
-        tonic::include_file_descriptor_set!("info_descriptor");
-
-    pub(crate) const COMMAND_FILE_DESCRIPTOR_SET: &[u8] =
-        tonic::include_file_descriptor_set!("command_descriptor");
-}
-
-// ShardService is a struct that contains a reference to the ShardManager
-#[derive(Debug)]
-struct ShardService {
-    // shard_manager is an Arc<ShardManager> that manages the shards
-    pub shard_manager: Arc<ShardManager>,
-}
-
-// Shard is a trait that defines the methods for the shard service
-#[tonic::async_trait]
-impl Shard for ShardService {
-    // shard_count is an async function that returns the count of shards and their ids
-    // It takes a Request<proto::ShardCountRequest> as a parameter and returns a Result<Response<proto::ShardCountResponse>, Status>
-    async fn shard_count(
-        &self,
-        _request: Request<proto::ShardCountRequest>,
-    ) -> Result<Response<proto::ShardCountResponse>, Status> {
-        trace!("Got a shard count request");
-        // Clone the shard_manager
-        let shard_manager = self.shard_manager.clone();
-        // Initialize an empty vector for the shard ids
-        let mut shard_ids = Vec::new();
-        // Iterate over the shard runners and push their ids to the shard_ids vector
-        for (shard_id, _) in shard_manager.runners.lock().await.iter() {
-            shard_ids.push(shard_id.0 as i32);
-        }
-        // Create a ShardCountResponse with the count of shards and their ids
-        let reply = proto::ShardCountResponse {
-            count: shard_ids.len() as i32,
-            shard_ids,
-        };
-        trace!("Completed a shard count request");
-        // Return the ShardCountResponse
-        Ok(Response::new(reply))
-    }
-
-    // shard_info is an async function that returns the information of a specific shard
-    // It takes a Request<proto::ShardInfoRequest> as a parameter and returns a Result<Response<proto::ShardInfoResponse>, Status>
-    async fn shard_info(
-        &self,
-        request: Request<proto::ShardInfoRequest>,
-    ) -> Result<Response<proto::ShardInfoResponse>, Status> {
-        trace!("Got a shard info request");
-        // Get the data from the request
-        let data = request.into_inner();
-        // Get the id of the shard
-        let id = data.shard_id;
-        // Clone the shard_manager
-        let shard_manager = self.shard_manager.clone();
-        // Lock the shard runners
-        let runners = shard_manager.runners.lock().await;
-        // If the shard is not found, return an error
-        if !runners.contains_key(&ShardId(id as u32)) {
-            return Err(Status::not_found("Shard not found"));
-        }
-        // Get the shard
-        let shard = runners.get(&ShardId(id as u32)).unwrap();
-        // Create a ShardInfoResponse with the shard id, latency, and stage
-        let reply = proto::ShardInfoResponse {
-            shard_id: id,
-            latency: shard.latency.unwrap_or_default().as_millis().to_string(),
-            stage: shard.stage.to_string(),
-        };
-        trace!("Completed a shard info request");
-        // Return the ShardInfoResponse
-        Ok(Response::new(reply))
-    }
-}
-
-pub struct InfoService {
-    pub bot_info: Arc<CurrentApplicationInfo>,
-    pub sys: Arc<RwLock<System>>,
-    pub os_info: Arc<os_info::Info>,
-}
-
-#[tonic::async_trait]
-impl Info for InfoService {
-    async fn get_info(
-        &self,
-        _request: Request<InfoRequest>,
-    ) -> Result<Response<InfoResponse>, Status> {
-        trace!("Got a info request");
-        let bot_info = self.bot_info.clone();
-        let sys = self.sys.clone();
-        let info = self.os_info.clone();
-        match sys.write() {
-            Ok(mut guard) => guard.refresh_all(),
-            _ => {}
-        }
-
-        let sys = sys.read().unwrap();
-        let processes = sys.processes();
-        let pid = match sysinfo::get_current_pid() {
-            Ok(pid) => pid,
-            _ => return Err(Status::internal("Process not found.")),
-        };
-        let process = match processes.get(&pid) {
-            Some(proc) => proc,
-            _ => return Err(Status::internal("Failed to get the process.")),
-        };
-
-        // system info
-        let os = format!(
-            "{}, {} {} {} {} {}",
-            info.os_type(),
-            info.bitness(),
-            info.version(),
-            info.codename().unwrap_or_default(),
-            info.architecture().unwrap_or_default(),
-            info.edition().unwrap_or_default()
-        );
-        let system_total_memory = format!("{}Gb", sys.total_memory() / 1024 / 1024 / 1024);
-        let system_used_memory = format!("{}Gb", sys.used_memory() / 1024 / 1024 / 1024);
-        let system_cpu_usage = format!("{}%", sys.global_cpu_info().cpu_usage());
-
-        let app_cpu = format!("{}%", process.cpu_usage());
-        let app_memory = process.memory();
-        let app_memory = format!("{:.2}Mb", app_memory / 1024 / 1024);
-
-        // bot info
-        let bot_name = bot_info.name.clone();
-        let version = APP_VERSION.to_string();
-        let bot_id = bot_info.id.to_string();
-        let bot_owner = match bot_info.owner.clone() {
-            Some(owner) => owner.name,
-            _ => return Err(Status::internal("Failed to get the bot owner.")),
-        };
-        let bot_activity = ACTIVITY_NAME.to_string();
-        let description = bot_info.description.clone();
-        let uptime = process.run_time();
-        let bot_uptime = format!("{}s", uptime);
-
-        let bot_info_data = BotInfoData {
-            bot_name,
-            version,
-            bot_uptime,
-            bot_id,
-            bot_owner,
-            bot_activity,
-            description,
-        };
-
-        let sys_info_data = SystemInfoData {
-            app_cpu,
-            app_memory,
-            os,
-            system_total_memory,
-            system_used_memory,
-            system_cpu_usage,
-        };
-
-        let info_response = InfoResponse {
-            bot_info: Option::from(bot_info_data),
-            sys_info: Option::from(sys_info_data),
-        };
-        trace!("Completed a info request");
-        Ok(Response::new(info_response))
-    }
-}
-
-pub struct CommandServices {
-    pub command_list: Arc<Vec<CommandItem>>,
-}
-
-#[tonic::async_trait]
-impl CommandService for CommandServices {
-    async fn command_list(
-        &self,
-        _request: Request<CommandListRequest>,
-    ) -> Result<Response<CommandListResponse>, Status> {
-        let cmd_list = &self.command_list.clone();
-        let cm_count = cmd_list.len();
-        let mut commands = Vec::new();
-        let mut sub_commands = Vec::new();
-        let mut sub_command_groups = Vec::new();
-        for cmd in cmd_list.iter() {
-            match cmd {
-                CommandItem::Command(c) => {
-                    commands.push(c.into());
-                }
-                CommandItem::Subcommand(s) => {
-                    sub_commands.push(s.into());
-                }
-                CommandItem::SubcommandGroup(sg) => {
-                    sub_command_groups.push(sg.into());
-                }
-            }
-        }
-        let response = CommandListResponse {
-            command_count: cm_count as i64,
-            commands,
-            sub_commands,
-            sub_command_groups,
-        };
-        Ok(Response::new(response))
-    }
-}
 /// `grpc_server_launcher` is an asynchronous function that launches the gRPC server for the shard service.
 /// It takes a reference to an `Arc<ShardManager>` as a parameter.
 /// It does not return a value.
@@ -247,7 +29,7 @@ impl CommandService for CommandServices {
 ///
 /// This function will panic if it fails to build the reflection service or if it fails to serve the gRPC server.
 pub async fn grpc_server_launcher(shard_manager: &Arc<ShardManager>) {
-    get_command_list();
+    get_list_of_all_command();
     // Clone the Arc<ShardManager>
     let shard_manager_arc: Arc<ShardManager> = shard_manager.clone();
 
@@ -272,9 +54,9 @@ pub async fn grpc_server_launcher(shard_manager: &Arc<ShardManager>) {
 
     // Configure the reflection service and register the file descriptor set for the shard service
     let reflection = tonic_reflection::server::Builder::configure()
-        .register_encoded_file_descriptor_set(proto::SHARD_FILE_DESCRIPTOR_SET)
-        .register_encoded_file_descriptor_set(proto::INFO_FILE_DESCRIPTOR_SET)
-        .register_encoded_file_descriptor_set(proto::COMMAND_FILE_DESCRIPTOR_SET)
+        .register_encoded_file_descriptor_set(service::shard::proto::SHARD_FILE_DESCRIPTOR_SET)
+        .register_encoded_file_descriptor_set(service::info::proto::INFO_FILE_DESCRIPTOR_SET)
+        .register_encoded_file_descriptor_set(service::command::proto::COMMAND_FILE_DESCRIPTOR_SET)
         .build()
         .unwrap();
 
@@ -299,9 +81,9 @@ pub async fn grpc_server_launcher(shard_manager: &Arc<ShardManager>) {
         tonic::transport::Server::builder()
             .tls_config(tls_config)
             .unwrap()
-            .add_service(ShardServer::new(shard_service))
-            .add_service(InfoServer::new(info_service))
-            .add_service(CommandServiceServer::new(command_service))
+            .add_service(get_shard_server(shard_service))
+            .add_service(get_info_server(info_service))
+            .add_service(get_command_server(command_service))
             .add_service(reflection)
             .serve(addr.parse().unwrap())
             .await
@@ -309,9 +91,9 @@ pub async fn grpc_server_launcher(shard_manager: &Arc<ShardManager>) {
     } else {
         // Build the gRPC server, add the ShardService and the reflection service, and serve the gRPC server
         tonic::transport::Server::builder()
-            .add_service(ShardServer::new(shard_service))
-            .add_service(InfoServer::new(info_service))
-            .add_service(CommandServiceServer::new(command_service))
+            .add_service(get_shard_server(shard_service))
+            .add_service(get_info_server(info_service))
+            .add_service(get_command_server(command_service))
             .add_service(reflection)
             .serve(addr.parse().unwrap())
             .await
@@ -349,66 +131,4 @@ fn generate_key() {
 
     std::fs::write(private_key_path, private_key).unwrap();
     std::fs::write(cert_path, certificate).unwrap();
-}
-
-impl From<&Command> for proto::Command {
-    fn from(command: &Command) -> Self {
-        proto::Command {
-            name: command.name.clone(),
-            description: command.desc.clone(),
-            args: command
-                .args
-                .clone()
-                .into_iter()
-                .map(|arg| (&arg).into())
-                .collect(),
-        }
-    }
-}
-
-impl From<&Arg> for proto::Arg {
-    fn from(arg: &Arg) -> Self {
-        proto::Arg {
-            name: arg.name.clone(),
-            description: arg.desc.clone(),
-            required: arg.required,
-            choices: arg.choices.clone(),
-        }
-    }
-}
-
-impl From<&SubCommand> for proto::SubCommand {
-    fn from(subcommand: &SubCommand) -> Self {
-        proto::SubCommand {
-            name: subcommand.name.clone(),
-            description: subcommand.desc.clone(),
-            commands: subcommand
-                .commands
-                .clone()
-                .into_iter()
-                .map(|commands| (&commands).into())
-                .collect(),
-        }
-    }
-}
-
-impl From<&SubCommandGroup> for proto::SubCommandGroup {
-    fn from(subcommand_group: &SubCommandGroup) -> Self {
-        proto::SubCommandGroup {
-            name: subcommand_group.name.clone(),
-            description: subcommand_group.desc.clone(),
-            sub_commands: subcommand_group
-                .subcommands
-                .clone()
-                .into_iter()
-                .map(|subcommands| (&subcommands).into())
-                .collect(),
-            commands: subcommand_group
-                .commands
-                .clone()
-                .into_iter()
-                .map(|commands| (&commands).into())
-                .collect(),
-        }
-    }
 }
