+++ src/grpc_server/mod.rs
@@ -1,2 +1,3 @@
 pub mod command_list;
 pub mod launcher;
+pub mod service;
