+++ src/grpc_server/service/mod.rs
@@ -1,0 +1,3 @@
+pub mod shard;
+pub mod info;
+pub mod command;
